Refactor: migrate ABG and HBG examples to tests/st
- Move aicpu_build_graph and host_build_graph cases from examples/
  to tests/st/, converting golden.py/kernel_config.py entries to
  SceneTestCase format
- Delete obsolete golden.py, kernel_config.py, README, and docs
- Add small-tile dispatch (<16,16,...>) to HBG paged attention
  kernels so the small1 case (head_dim=16, block_size=16) no longer
  segfaults from buffer overruns

+++ tests/st/a2a3/aicpu_build_graph/paged_attention_unroll/kernels/orchestration/paged_attention_orch.cpp
@@ -112,7 +112,7 @@ aicpu_orchestration_entry(PTO2Runtime *rt, const ChipStorageTaskArgs &orch_args)
     uint64_t scale_value = orch_args.scalar(0);
 
     uint64_t q_head_num = num_heads;
-    uint64_t q_tile = std::min(num_heads, 128UL);
+    uint64_t q_tile = std::min(num_heads, static_cast<uint64_t>(128));
     uint64_t q_loop = (q_head_num + q_tile - 1) / q_tile;
     CYCLE_COUNT_LAP(prof_param_extract);
